refactor: Criando um arquivo utils para armazenar funções em comum do sistema.
