+++ src/components/PopUp.vue
@@ -21,6 +21,7 @@
 
 <script setup lang="ts">
 import { type LatestEquipmentInfo } from "@/types/types"
+import { formatDate } from "@/utils/utils"
 
 const props = defineProps({
     position: {
@@ -28,19 +29,6 @@ const props = defineProps({
         required: true,
     },
 });
-
-const formatDate = (dateString: string) => {
-    const date = new Date(dateString);
-
-    const day = String(date.getDate()).padStart(2, "0");
-    const month = String(date.getMonth() + 1).padStart(2, "0");
-    const year = date.getFullYear();
-
-    const hours = String(date.getHours()).padStart(2, "0");
-    const minutes = String(date.getMinutes()).padStart(2, "0");
-
-    return `${day}/${month}/${year} às ${hours}:${minutes}`;
-};
 </script>
 
 <style scoped>
